Cleanups * Moved instance storage to unordered_map. * Added back an instance counter to avoid overlapping instance IDs. * Cleaned up the CSV2YAML tool to not export the default values for TimeLimit and IdleTimeOut.

+++ src/map/map.hpp
@@ -753,7 +753,7 @@ struct map_data {
 	int mob_delete_timer;	// Timer ID for map_removemobs_timer [Skotlex]
 
 	// Instance Variables
-	unsigned short instance_id;
+	int instance_id;
 	int instance_src_map;
 
 	/* rAthena Local Chat */
